fix(getAdminAddFields): fix endless-loop
Fixes endless-loop with page-refreshes on admin -> add service

Current versions of blesta are triggering the .change() event on all fields at page load in some circumstances (once a pricing is selected).
Because the ispapi plugin is listening for change events in the domain field and then refreshing the page, this resulted in an endless-loop (page always refreshing)
This fix will prevent a page refresh if the domain field value has not actually changed.